+++ packages/current-time/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @vuecalendar/current-time
 
+## 1.0.7
+
+### Patch Changes
+
+- Fix registry
+- Updated dependencies
+  - @vuecalendar/core@1.0.7
+
+## 1.0.6
+
+### Patch Changes
+
+- Fix registry
+- Fix registry
+- Fix registry
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @vuecalendar/core@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- Fix registry
+- Updated dependencies
+  - @vuecalendar/core@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Fix registry 2
+- Fix registry 3
+- Fix registry 3
+- Fix registry 4
+- Fix registry 4
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @vuecalendar/core@1.0.4
+
+## 1.0.3
+
+### Patch Changes
+
+- Fix registry
+- Updated dependencies
+  - @vuecalendar/core@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Describe the premium fix
+- Updated dependencies
+  - @vuecalendar/core@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes